Extract function related functions (#9721)
